Making google_dialogflow_cx_version update work (#9177)
* Making google_dialogflow_cx_version update work

I also took the opportunity to turn the Version and Environment resources into mmv1 resources instead of handwritten ones. It needed only a small update to the dialogflow async operations code.

* Oops, was updating an immutable field

* Skipping documentation generation for "regional" tests

It's confusing and not useful, given that the "regional" tests were specifically to exercise a bug.

* Reverting possibly breaking cleanup

* Preserving CustomizeDiff behavior

* PR feedback: switching location to us-central1 for tests + making issue name more parseable
[upstream:33c1fb4639fbad7163fd43820605d89f7858b699]
