+++ google/provider/provider.go
@@ -997,9 +997,9 @@ func DatasourceMapWithErrors() (map[string]*schema.Resource, error) {
 		})
 }
 
-// Generated resources: 329
+// Generated resources: 331
 // Generated IAM resources: 207
-// Total generated resources: 536
+// Total generated resources: 538
 func ResourceMap() map[string]*schema.Resource {
 	resourceMap, _ := ResourceMapWithErrors()
 	return resourceMap
@@ -1301,11 +1301,13 @@ func ResourceMapWithErrors() (map[string]*schema.Resource, error) {
 			"google_dialogflow_intent":                                       dialogflow.ResourceDialogflowIntent(),
 			"google_dialogflow_cx_agent":                                     dialogflowcx.ResourceDialogflowCXAgent(),
 			"google_dialogflow_cx_entity_type":                               dialogflowcx.ResourceDialogflowCXEntityType(),
+			"google_dialogflow_cx_environment":                               dialogflowcx.ResourceDialogflowCXEnvironment(),
 			"google_dialogflow_cx_flow":                                      dialogflowcx.ResourceDialogflowCXFlow(),
 			"google_dialogflow_cx_intent":                                    dialogflowcx.ResourceDialogflowCXIntent(),
 			"google_dialogflow_cx_page":                                      dialogflowcx.ResourceDialogflowCXPage(),
 			"google_dialogflow_cx_security_settings":                         dialogflowcx.ResourceDialogflowCXSecuritySettings(),
 			"google_dialogflow_cx_test_case":                                 dialogflowcx.ResourceDialogflowCXTestCase(),
+			"google_dialogflow_cx_version":                                   dialogflowcx.ResourceDialogflowCXVersion(),
 			"google_dialogflow_cx_webhook":                                   dialogflowcx.ResourceDialogflowCXWebhook(),
 			"google_dns_managed_zone":                                        dns.ResourceDNSManagedZone(),
 			"google_dns_managed_zone_iam_binding":                            tpgiamresource.ResourceIamBinding(dns.DNSManagedZoneIamSchema, dns.DNSManagedZoneIamUpdaterProducer, dns.DNSManagedZoneIdParseFunc),
@@ -1569,8 +1571,6 @@ func ResourceMapWithErrors() (map[string]*schema.Resource, error) {
 			"google_dataflow_job":                           dataflow.ResourceDataflowJob(),
 			"google_dataproc_cluster":                       dataproc.ResourceDataprocCluster(),
 			"google_dataproc_job":                           dataproc.ResourceDataprocJob(),
-			"google_dialogflow_cx_version":                  dialogflowcx.ResourceDialogflowCXVersion(),
-			"google_dialogflow_cx_environment":              dialogflowcx.ResourceDialogflowCXEnvironment(),
 			"google_dns_record_set":                         dns.ResourceDnsRecordSet(),
 			"google_endpoints_service":                      servicemanagement.ResourceEndpointsService(),
 			"google_folder":                                 resourcemanager.ResourceGoogleFolder(),
