+++ google/services/dialogflowcx/dialogflow_cx_operation.go
@@ -5,6 +5,7 @@ package dialogflowcx
 import (
 	"encoding/json"
 	"fmt"
+	"regexp"
 	"time"
 
 	"github.com/hashicorp/terraform-provider-google/google/tpgresource"
@@ -15,7 +16,6 @@ type DialogflowCXOperationWaiter struct {
 	Config    *transport_tpg.Config
 	UserAgent string
 	tpgresource.CommonOperationWaiter
-	Location string
 }
 
 func (w *DialogflowCXOperationWaiter) QueryOp() (interface{}, error) {
@@ -23,7 +23,18 @@ func (w *DialogflowCXOperationWaiter) QueryOp() (interface{}, error) {
 		return nil, fmt.Errorf("Cannot query operation, it's unset or nil.")
 	}
 	// Returns the proper get.
-	url := fmt.Sprintf("https://%s-dialogflow.googleapis.com/v3/%s", w.Location, w.CommonOperationWaiter.Op.Name)
+	location := ""
+	if parts := regexp.MustCompile(`locations\/([^\/]*)\/`).FindStringSubmatch(w.CommonOperationWaiter.Op.Name); parts != nil {
+		location = parts[1]
+	} else {
+		return nil, fmt.Errorf(
+			"Saw %s when the op name is expected to contains location %s",
+			w.CommonOperationWaiter.Op.Name,
+			"projects/{{project}}/locations/{{location}}/...",
+		)
+	}
+
+	url := fmt.Sprintf("https://%s-dialogflow.googleapis.com/v3/%s", location, w.CommonOperationWaiter.Op.Name)
 
 	return transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
 		Config:    w.Config,
@@ -33,11 +44,10 @@ func (w *DialogflowCXOperationWaiter) QueryOp() (interface{}, error) {
 	})
 }
 
-func createDialogflowCXWaiter(config *transport_tpg.Config, op map[string]interface{}, activity, userAgent, location string) (*DialogflowCXOperationWaiter, error) {
+func createDialogflowCXWaiter(config *transport_tpg.Config, op map[string]interface{}, activity, userAgent string) (*DialogflowCXOperationWaiter, error) {
 	w := &DialogflowCXOperationWaiter{
 		Config:    config,
 		UserAgent: userAgent,
-		Location:  location,
 	}
 	if err := w.CommonOperationWaiter.SetOp(op); err != nil {
 		return nil, err
@@ -46,8 +56,8 @@ func createDialogflowCXWaiter(config *transport_tpg.Config, op map[string]interf
 }
 
 // nolint: deadcode,unused
-func DialogflowCXOperationWaitTimeWithResponse(config *transport_tpg.Config, op map[string]interface{}, response *map[string]interface{}, activity, userAgent, location string, timeout time.Duration) error {
-	w, err := createDialogflowCXWaiter(config, op, activity, userAgent, location)
+func DialogflowCXOperationWaitTimeWithResponse(config *transport_tpg.Config, op map[string]interface{}, response *map[string]interface{}, activity, userAgent string, timeout time.Duration) error {
+	w, err := createDialogflowCXWaiter(config, op, activity, userAgent)
 	if err != nil {
 		return err
 	}
@@ -57,12 +67,12 @@ func DialogflowCXOperationWaitTimeWithResponse(config *transport_tpg.Config, op 
 	return json.Unmarshal([]byte(w.CommonOperationWaiter.Op.Response), response)
 }
 
-func DialogflowCXOperationWaitTime(config *transport_tpg.Config, op map[string]interface{}, activity, userAgent, location string, timeout time.Duration) error {
+func DialogflowCXOperationWaitTime(config *transport_tpg.Config, op map[string]interface{}, activity, userAgent string, timeout time.Duration) error {
 	if val, ok := op["name"]; !ok || val == "" {
 		// This was a synchronous call - there is no operation to wait for.
 		return nil
 	}
-	w, err := createDialogflowCXWaiter(config, op, activity, userAgent, location)
+	w, err := createDialogflowCXWaiter(config, op, activity, userAgent)
 	if err != nil {
 		// If w is nil, the op was synchronous.
 		return err
