+++ google/services/dialogflowcx/resource_dialogflow_cx_environment.go
@@ -1,5 +1,20 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
+
+// ----------------------------------------------------------------------------
+//
+//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
+//
+// ----------------------------------------------------------------------------
+//
+//     This file is automatically generated by Magic Modules and manual
+//     changes will be clobbered when the file is regenerated.
+//
+//     Please read more about how to change this file in
+//     .github/CONTRIBUTING.md.
+//
+// ----------------------------------------------------------------------------
+
 package dialogflowcx
 
 import (
@@ -166,7 +181,7 @@ func resourceDialogflowCXEnvironmentCreate(d *schema.ResourceData, meta interfac
 	// identity fields and d.Id() before read
 	var opRes map[string]interface{}
 	err = DialogflowCXOperationWaitTimeWithResponse(
-		config, res, &opRes, "Creating Environment", userAgent, location,
+		config, res, &opRes, "Creating Environment", userAgent,
 		d.Timeout(schema.TimeoutCreate))
 	if err != nil {
 		// The resource didn't actually create
@@ -329,28 +344,31 @@ func resourceDialogflowCXEnvironmentUpdate(d *schema.ResourceData, meta interfac
 		billingProject = bp
 	}
 
-	res, err := transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
-		Config:    config,
-		Method:    "PATCH",
-		Project:   billingProject,
-		RawURL:    url,
-		UserAgent: userAgent,
-		Body:      obj,
-		Timeout:   d.Timeout(schema.TimeoutUpdate),
-	})
+	// if updateMask is empty we are not updating anything so skip the post
+	if len(updateMask) > 0 {
+		res, err := transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
+			Config:    config,
+			Method:    "PATCH",
+			Project:   billingProject,
+			RawURL:    url,
+			UserAgent: userAgent,
+			Body:      obj,
+			Timeout:   d.Timeout(schema.TimeoutUpdate),
+		})
 
-	if err != nil {
-		return fmt.Errorf("Error updating Environment %q: %s", d.Id(), err)
-	} else {
-		log.Printf("[DEBUG] Finished updating Environment %q: %#v", d.Id(), res)
-	}
+		if err != nil {
+			return fmt.Errorf("Error updating Environment %q: %s", d.Id(), err)
+		} else {
+			log.Printf("[DEBUG] Finished updating Environment %q: %#v", d.Id(), res)
+		}
 
-	err = DialogflowCXOperationWaitTime(
-		config, res, "Updating Environment", userAgent, location,
-		d.Timeout(schema.TimeoutUpdate))
+		err = DialogflowCXOperationWaitTime(
+			config, res, "Updating Environment", userAgent,
+			d.Timeout(schema.TimeoutUpdate))
 
-	if err != nil {
-		return err
+		if err != nil {
+			return err
+		}
 	}
 
 	return resourceDialogflowCXEnvironmentRead(d, meta)
@@ -404,14 +422,6 @@ func resourceDialogflowCXEnvironmentDelete(d *schema.ResourceData, meta interfac
 	})
 	if err != nil {
 		return transport_tpg.HandleNotFoundError(err, d, "Environment")
-	}
-
-	err = DialogflowCXOperationWaitTime(
-		config, res, "Deleting Environment", userAgent, location,
-		d.Timeout(schema.TimeoutDelete))
-
-	if err != nil {
-		return err
 	}
 
 	log.Printf("[DEBUG] Finished deleting Environment %q: %#v", d.Id(), res)
