+++ google/services/dialogflowcx/resource_dialogflow_cx_version.go
@@ -1,5 +1,20 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
+
+// ----------------------------------------------------------------------------
+//
+//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
+//
+// ----------------------------------------------------------------------------
+//
+//     This file is automatically generated by Magic Modules and manual
+//     changes will be clobbered when the file is regenerated.
+//
+//     Please read more about how to change this file in
+//     .github/CONTRIBUTING.md.
+//
+// ----------------------------------------------------------------------------
+
 package dialogflowcx
 
 import (
@@ -186,7 +201,7 @@ func resourceDialogflowCXVersionCreate(d *schema.ResourceData, meta interface{})
 	// identity fields and d.Id() before read
 	var opRes map[string]interface{}
 	err = DialogflowCXOperationWaitTimeWithResponse(
-		config, res, &opRes, "Creating Version", userAgent, location,
+		config, res, &opRes, "Creating Version", userAgent,
 		d.Timeout(schema.TimeoutCreate))
 	if err != nil {
 		// The resource didn't actually create
@@ -342,28 +357,24 @@ func resourceDialogflowCXVersionUpdate(d *schema.ResourceData, meta interface{})
 		billingProject = bp
 	}
 
-	res, err := transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
-		Config:    config,
-		Method:    "PATCH",
-		Project:   billingProject,
-		RawURL:    url,
-		UserAgent: userAgent,
-		Body:      obj,
-		Timeout:   d.Timeout(schema.TimeoutUpdate),
-	})
+	// if updateMask is empty we are not updating anything so skip the post
+	if len(updateMask) > 0 {
+		res, err := transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
+			Config:    config,
+			Method:    "PATCH",
+			Project:   billingProject,
+			RawURL:    url,
+			UserAgent: userAgent,
+			Body:      obj,
+			Timeout:   d.Timeout(schema.TimeoutUpdate),
+		})
 
-	if err != nil {
-		return fmt.Errorf("Error updating Version %q: %s", d.Id(), err)
-	} else {
-		log.Printf("[DEBUG] Finished updating Version %q: %#v", d.Id(), res)
-	}
+		if err != nil {
+			return fmt.Errorf("Error updating Version %q: %s", d.Id(), err)
+		} else {
+			log.Printf("[DEBUG] Finished updating Version %q: %#v", d.Id(), res)
+		}
 
-	err = DialogflowCXOperationWaitTime(
-		config, res, "Updating Version", userAgent, location,
-		d.Timeout(schema.TimeoutUpdate))
-
-	if err != nil {
-		return err
 	}
 
 	return resourceDialogflowCXVersionRead(d, meta)
@@ -417,14 +428,6 @@ func resourceDialogflowCXVersionDelete(d *schema.ResourceData, meta interface{})
 	})
 	if err != nil {
 		return transport_tpg.HandleNotFoundError(err, d, "Version")
-	}
-
-	err = DialogflowCXOperationWaitTime(
-		config, res, "Deleting Version", userAgent, location,
-		d.Timeout(schema.TimeoutDelete))
-
-	if err != nil {
-		return err
 	}
 
 	log.Printf("[DEBUG] Finished deleting Version %q: %#v", d.Id(), res)
